+++ ModelLabs/CommonCloud/AzureStorage/Entities/TotalProduction.cs
@@ -15,19 +15,19 @@ namespace CommonCloud.AzureStorage.Entities
         [DataMember]
         private int _id;
         [DataMember]
-        private float _totalGeneration;
+        private double _totalGeneration;
         [DataMember]
-        private float _cO2Reduction;
+        private double _cO2Reduction;
         [DataMember]
-        private float _cO2Emission;
+        private double _cO2Emission;
         [DataMember]
-        private float _totalCost;
+        private double _totalCost;
         [DataMember]
-        private float _profit;
+        private double _profit;
         [DataMember]
         private DateTime _timeOfCalculation;
 
-        public TotalProduction(float totalGeneration, float cO2Reduction, float cO2Emission, float totalCost, float profit, DateTime timeOfCalculation)
+        public TotalProduction(double totalGeneration, double cO2Reduction, double cO2Emission, double totalCost, double profit, DateTime timeOfCalculation)
         {
             Id = 1;
             TotalGeneration = totalGeneration;
@@ -44,11 +44,11 @@ namespace CommonCloud.AzureStorage.Entities
         public TotalProduction()
         { }
         public int Id { get => _id; set => _id = value; }
-        public float TotalGeneration { get => _totalGeneration; set => _totalGeneration = value; }
-        public float CO2Reduction { get => _cO2Reduction; set => _cO2Reduction = value; }
-        public float CO2Emission { get => _cO2Emission; set => _cO2Emission = value; }
-        public float TotalCost { get => _totalCost; set => _totalCost = value; }
-        public float Profit { get => _profit; set => _profit = value; }
+        public double TotalGeneration { get => _totalGeneration; set => _totalGeneration = value; }
+        public double CO2Reduction { get => _cO2Reduction; set => _cO2Reduction = value; }
+        public double CO2Emission { get => _cO2Emission; set => _cO2Emission = value; }
+        public double TotalCost { get => _totalCost; set => _totalCost = value; }
+        public double Profit { get => _profit; set => _profit = value; }
         public DateTime TimeOfCalculation { get => _timeOfCalculation; set => _timeOfCalculation = value; }
